"Updated two-way serial, minor int type changes for speed function arguments, added confirmResponse check & CRC gen."

+++ Marlin/src/api/tools/tools.h
@@ -39,5 +39,5 @@ int setDispenseHeight(Tool tool, float height);
 float getDispenseHeight(Tool tool);
 
 // Router
-int setRotationSpeed(Tool tool, float speed);
+int setRotationSpeed(Tool tool, int speed);
 float getRotationSpeed(Tool tool);
